Use Record over {}
