+++ lib/components/resolver.js
@@ -2,12 +2,13 @@
 
 const util = require('../util')
 // eslint-disable-next-line no-unused-vars
-const { Buffer, SourceFile, Path, Library, baseDefinitions } = require('../file')
+const { Buffer, SourceFile, Path, Library } = require('../file')
 const { deepRequire, createToManyAssociation, createToOneAssociation, createArrayOf, createCompositionOfMany, createCompositionOfOne } = require('./wrappers')
 const { StructuredInlineDeclarationResolver } = require('./inline')
 const { isInlineEnumType, propertyToInlineEnumName } = require('./enum')
 const { isReferenceType } = require('./reference')
 const { isEntity } = require('../csn')
+const { baseDefinitions } = require('./basedefs')
 
 /** @typedef {{ cardinality?: { max?: '*' | number } }} EntityCSN */
 /** @typedef {{ definitions?: Object<string, EntityCSN> }} CSN */
