+++ lib/components/typescript.js
@@ -1,0 +1,3 @@
+module.exports = {
+    empty: 'Record<never, never>'
+}
