+++ lib/components/wrappers.js
@@ -1,6 +1,7 @@
 'use strict'
 
-const base = require('../file').baseDefinitions.path.asIdentifier()
+// this was derived from baseDefinitions before, but caused a circular dependency
+const base = '__'
 
 /**
  * Wraps type into association to scalar.
@@ -38,6 +39,13 @@ const createCompositionOfMany = t => `${base}.Composition.of.many<${t}>`
 const createArrayOf = t => `Array<${t}>`
 
 /**
+ * Wraps type into object braces
+ * @param {string} t the properties, stringified and comma separated.
+ * @returns {string}
+ */
+const createObjectOf = t => `{${t}}`
+
+/**
  * Wraps type into a deep require (removes all posibilities of undefined recursively).
  * @param {string} t the singular type name. 
  * @param {string?} lookup a property lookup of the required type (`['Foo']`)
@@ -57,6 +65,7 @@ const docify = doc => doc
 
 module.exports = {
     createArrayOf,
+    createObjectOf,
     createToOneAssociation,
     createToManyAssociation,
     createCompositionOfOne,
