+++ lib/file.js
@@ -1,10 +1,12 @@
 'use strict'
 
 const fs = require('fs').promises
+const path = require('path')
 const { readFileSync } = require('fs')
 const { printEnum, propertyToInlineEnumName, stringifyEnumImplementation } = require('./components/enum')
 const { normalise } = require('./components/identifier')
-const path = require('path')
+const { empty } = require('./components/typescript')
+const { createObjectOf } = require('./components/wrappers')
 
 const AUTO_GEN_NOTE = '// This is an automatically generated file. Please do not change its contents manually!'
 
@@ -156,6 +158,9 @@ class SourceFile extends File {
      */
     static stringifyLambda({name, parameters=[], returns='any', initialiser, isStatic=false, kind}) {
         const parameterTypes = parameters.map(([n, t]) => `${normalise(n)}: ${t}`).join(', ')
+        const parameterTypeAsObject = parameterTypes.length
+            ? createObjectOf(parameterTypes)
+            : empty
         const callableSignature = `(${parameterTypes}): ${returns}`
         let prefix = name ? `${normalise(name)}: `: ''
         if (prefix && isStatic) {
@@ -163,7 +168,7 @@ class SourceFile extends File {
         }
         const kindDef = kind ? `, kind: '${kind}'` : ''
         const suffix = initialiser ? ` = ${initialiser}` : ''
-        const lambda = `{ ${callableSignature}, __parameters: {${parameterTypes}}, __returns: ${returns}${kindDef}}`
+        const lambda = `{ ${callableSignature}, __parameters: ${parameterTypeAsObject}, __returns: ${returns}${kindDef}}`
         return prefix + lambda + suffix
     }
 
@@ -599,67 +604,6 @@ class FileRepository {
     }
 }
 
-// eslint-disable-next-line no-template-curly-in-string
-const dateRegex = '`${number}${number}${number}${number}-${number}${number}-${number}${number}`'
-// eslint-disable-next-line no-template-curly-in-string
-const timeRegex = '`${number}${number}:${number}${number}:${number}${number}`'
-
-/**
- * Base definitions used throughout the typing process,
- * such as Associations and Compositions.
- * @type {SourceFile}
- */
-const baseDefinitions = new SourceFile('_')
-// FIXME: this should be a library someday
-baseDefinitions.addPreamble(`
-export namespace Association {
-    export type to <T> = T;
-    export namespace to {
-        export type many <T extends readonly any[]> = T;
-    }
-}
-
-export namespace Composition {
-    export type of <T> = T;
-    export namespace of {
-        export type many <T extends readonly any[]> = T;
-    }
-}
-
-export class Entity {
-    static data<T extends Entity> (this:T, _input:Object) : T {
-        return {} as T // mock
-    }
-}
-
-export type EntitySet<T> = T[] & {
-    data (input:object[]) : T[]
-    data (input:object) : T
-};
-
-export type DeepRequired<T> = { 
-    [K in keyof T]: DeepRequired<T[K]>
-} & Exclude<Required<T>, null>;
-
-
-/**
- * Dates and timestamps are strings during runtime, so cds-typer represents them as such.
- */
-export type CdsDate = ${dateRegex};
-/**
- * @see {@link CdsDate}
- */
-export type CdsDateTime = string;
-/**
- * @see {@link CdsDate}
- */
-export type CdsTime = ${timeRegex};
-/**
- * @see {@link CdsDate}
- */
-export type CdsTimestamp = string;
-`)
-
 /**
  * Writes the files to disk. For each source, a index.d.ts holding the type definitions
  * and a index.js holding implementation stubs is generated at the appropriate directory.
@@ -696,6 +640,5 @@ module.exports = {
     FileRepository,
     SourceFile,
     Path,
-    writeout,
-    baseDefinitions,
+    writeout
 }
