+++ lib/visitor.js
@@ -4,13 +4,15 @@ const util = require('./util')
 
 const { amendCSN, isView, isUnresolved, propagateForeignKeys, isDraftEnabled, isType, isProjection } = require('./csn')
 // eslint-disable-next-line no-unused-vars
-const { SourceFile, FileRepository, baseDefinitions, Buffer } = require('./file')
+const { SourceFile, FileRepository, Buffer } = require('./file')
 const { FlatInlineDeclarationResolver, StructuredInlineDeclarationResolver } = require('./components/inline')
 const { Resolver, resolveBuiltin } = require('./components/resolver')
 const { Logger } = require('./logging')
 const { docify } = require('./components/wrappers')
 const { csnToEnumPairs, propertyToInlineEnumName, isInlineEnumType, stringifyEnumType } = require('./components/enum')
 const { isReferenceType } = require('./components/reference')
+const { empty } = require('./components/typescript')
+const { baseDefinitions } = require('./components/basedefs')
 
 /** @typedef {import('./file').File} File */
 /** @typedef {{ entity: String }} Context */
@@ -201,14 +203,18 @@ class Visitor {
                         file.addInlineEnum(clean, name, e.name, csnToEnumPairs(e, {unwrapVals: true}))
                     }
                     const actions = Object.entries(entity.actions ?? {})
-                    buffer.addIndentedBlock('static actions: {', 
-                        actions.map(([aname, action]) => SourceFile.stringifyLambda({
-                            name: aname,
-                            parameters: this.#stringifyFunctionParams(action.params, file),
-                            returns: action.returns ? this.resolver.resolveAndRequire(action.returns, file).typeName : 'any',
-                            kind: action.kind
-                        }))
-                        , '}') // end of actions
+                    if (actions.length) {
+                        buffer.addIndentedBlock('static readonly actions: {', 
+                            actions.map(([aname, action]) => SourceFile.stringifyLambda({
+                                name: aname,
+                                parameters: this.#stringifyFunctionParams(action.params, file),
+                                returns: action.returns ? this.resolver.resolveAndRequire(action.returns, file).typeName : 'any',
+                                kind: action.kind
+                            }))
+                            , '}') // end of actions
+                    } else {
+                        buffer.add(`static readonly actions: ${empty}`)
+                    }
                 }.bind(this))
             }.bind(this), '};') // end of generated class
         }.bind(this), '}') // end of aspect
